Implement distribution metrics in the Sentry, with arbitrary number of fields.
Distribution metrics are well-suited for recording "events" and the time these
events take, for performance measurements. They bucket durations in buckets,
and keep track of the number of samples in each bucket. As this structure also
inherently keeps track of the *total* number of samples, it can be used as a
simple event counter as well, obviating the need for a counter metric next to
it counting the same thing.

In order to be compatible with the needs of the KVM platform to track events
that happen where new memory allocations would not be possible, the code for
adding a sample to a distribution is optimized to be fast and allocation-free.
The tradeoff there mostly comes in the form of memory, such as requiring a
weird new `fieldMapper` recursive struct that acts as a lookup table for the
concatenated key containing the values of all the fields for which the sample
is being recorded. Since we do not expect to deal with large number of field
combinations, this should not be a problem.

Another tradeoff this imposes is the lack of support for a generic `Bucketer`
interface allowing users to define their own bucketing scheme, as we would not
be able to enforce the lack of allocations in custom `Bucketer`
implementations, nor enforce `+checkescape` on them. However, since in
practice all bucketing implementations will probably reside in `metric.go`,
this is worked around by just having the distribution metric code refer to
`Bucketer` implementations as references and call them directly (without the
interface indirection). Since there is only one implementation currently
(`ExponentialBucketer`), this is faster than using the interface.

PiperOrigin-RevId: 435212668